chore(cve details page): Rename cvePage to CVEDetailsPage
