+++ src/Components/PresentationalComponents/CVEInfoBox/CVEInfoBox.js
@@ -1,7 +1,7 @@
 import { Bullseye, Split, SplitItem, Text, TextContent, TextVariants } from '@patternfly/react-core';
 import propTypes from 'prop-types';
 import React from 'react';
-import { CVEPageContext } from '../../SmartComponents/CVEPage/CVEPage';
+import { CVEPageContext } from '../../SmartComponents/CVEDetailsPage/CVEDetailsPage';
 import WithLoader from '../WithLoader/WithLoader';
 import './CVEInfoBox.scss';
