+++ src/Components/PresentationalComponents/CVEPageDetails/CVEPageDetails.js
@@ -1,7 +1,7 @@
 import { Grid, GridItem } from '@patternfly/react-core';
 import propTypes from 'prop-types';
 import React from 'react';
-import { CVEPageContext } from '../../SmartComponents/CVEPage/CVEPage';
+import { CVEPageContext } from '../../SmartComponents/CVEDetailsPage/CVEDetailsPage';
 import CVEPageDetailsDescription from '../CVEPageDetailsDescription/CVEPageDetailsDescription';
 import CVEPageDetailsSeverity from '../CVEPageDetailsSeverity/CVEPageDetailsSeverity';
 import WithLoader from '../WithLoader/WithLoader';
