+++ src/Components/PresentationalComponents/CvssVector/CvssVector.js
@@ -3,7 +3,7 @@ import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons';
 import { Table, TableBody, TableHeader, TableVariant } from '@patternfly/react-table';
 import propTypes from 'prop-types';
 import React from 'react';
-import { CVEPageContext } from '../../SmartComponents/CVEPage/CVEPage';
+import { CVEPageContext } from '../../SmartComponents/CVEDetailsPage/CVEDetailsPage';
 import WithLoader from '../WithLoader/WithLoader';
 import './CvssVector.scss';
 import { injectIntl } from 'react-intl';
