+++ src/Components/SmartComponents/CVEDetailsPage/CVEDetailsPage.js
@@ -23,7 +23,7 @@ import VulnerabilityHeader from '../VulnerabilityHeader/VulnerabilityHeader';
 
 export const CVEPageContext = React.createContext({ isLoading: true });
 
-class CVEPage extends React.Component {
+class CVEDetailsPage extends React.Component {
     static propTypes = {
         match: propTypes.object,
         cveDetails: propTypes.object,
@@ -160,5 +160,5 @@ export default withRouter(
     connect(
         mapStateToProps,
         mapDispatchToProps
-    )(CVEPage)
+    )(CVEDetailsPage)
 );
