+++ src/Helpers/constants.js
@@ -28,7 +28,7 @@ export const PATHS = {
         title: 'CVEs',
         to: '/cves'
     },
-    cvePage: {
+    cveDetailsPage: {
         title: 'CVE',
         to: '/cves/:cve'
     },
