+++ src/Utilities/Routes.js
@@ -5,16 +5,34 @@ import { Redirect, Route, Switch, useLocation } from 'react-router-dom';
 import asyncComponent from './asyncComponent';
 import { PATHS } from '../Helpers/constants';
 
-const CVEPage = asyncComponent(() => import(/* webpackChunkName: "CVEPage" */ '../Components/SmartComponents/CVEPage/CVEPage'));
-const Systems = asyncComponent(() => import(/* webpackChunkName: "Systems" */ '../Components/SmartComponents/Systems/Systems'));
+const Systems = asyncComponent(() =>
+    import(
+        /* webpackChunkName: "Systems" */ '../Components/SmartComponents/Systems/Systems'
+    )
+);
 
 const LandingPage = asyncComponent(() =>
-    import(/* webpackChunkName: "LandingPage" */ '../Components/SmartComponents/LandingPage/LandingPage'));
-const InventoryDetail = asyncComponent(() =>
-    import(/* webpackChunkName: "InventoryDetail" */ '../Components/SmartComponents/InventoryDetail/InventoryDetail')
+    import(
+        /* webpackChunkName: "LandingPage" */ '../Components/SmartComponents/LandingPage/LandingPage'
+    )
 );
+
 const Upgrade = asyncComponent(() =>
-    import(/* webpackChunkName: "Upgrade" */ '../Components/PresentationalComponents/StaticPages/UpgradePage')
+    import(
+        /* webpackChunkName: "Upgrade" */ '../Components/PresentationalComponents/StaticPages/UpgradePage'
+    )
+);
+
+const CVEDetailsPage = asyncComponent(() =>
+    import(
+        /* webpackChunkName: "CVEDetailsPage" */ '../Components/SmartComponents/CVEDetailsPage/CVEDetailsPage'
+    )
+);
+
+const InventoryDetail = asyncComponent(() =>
+    import(
+        /* webpackChunkName: "InventoryDetail" */ '../Components/SmartComponents/InventoryDetail/InventoryDetail'
+    )
 );
 
 const InsightsRoute = ({ component: Component, rootClass = 'vulnerability', ...rest }) => {
@@ -39,7 +57,7 @@ export const Routes = () => {
     return (
         <Switch>
             <Redirect from={PATHS.cvepagePath.to} to={PATHS.inventoryDetail.to} />
-            <InsightsRoute exact path={PATHS.cvePage.to} component={CVEPage} />
+            <InsightsRoute exact path={PATHS.cveDetailsPage.to} component={CVEDetailsPage} />
             <InsightsRoute exact path={PATHS.inventoryDetail.to}  component={InventoryDetail} />
             <InsightsRoute exact path={PATHS.cves.to} component={LandingPage} />
             <InsightsRoute exact path={PATHS.systems.to} component={Systems} />
